requirements.txt
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
dash==1.19.0
dash-auth==1.4.1
dash-bootstrap-components==0.11.1
dash-core-components==1.15.0
dash-daq==0.5.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.8.0
Flask-SeaSurf==0.3.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.0
pandas==1.2.1
plotly==4.14.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
ua-parser==0.10.0
urllib3==1.26.3
Werkzeug==1.0.1
